pyproject.toml
[build-system]
    build-backend = "uv_build"
    requires      = [ "uv_build>=0.8.6,<0.9.0" ]

[project]
    authors = [ { name = "Snaptraks" } ]
    classifiers = [
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Topic :: Communications :: Chat",
        "Topic :: Internet",
        "Topic :: Utilities",
    ]
    dependencies = [
        "aiosqlite>=0.21",
        "discord-py==2.6.3",
        "sqlalchemy[asyncio]>=2.0,<3",
    ]
    description = "A collection of Cogs for discord.py bots."
    license = { file = "LICENSE.txt" }
    name = "snapcogs"
    readme = { file = "README.md", content-type = "text/markdown" }
    requires-python = ">=3.12"
    version = "2025.09.08"

[project.urls]
    "Bug Reports" = "https://github.com/Snaptraks/SnapCogs/issues"
    Documentation = "https://github.com/Snaptraks/SnapCogs/blob/master/README.md"

[project.optional-dependencies]
    all = [
        "snapcogs[fun,horoscope,measurements,timestamps]",
    ]
    fun = [
        "pillow>=11.3.0",
    ]
    horoscope = [
        "beautifulsoup4>=4.9.3",
    ]
    measurements = [
        "pint>=0.19",
    ]
    timestamps = [
        "python-dateutil>=2.0",
        "pytz",
    ]
[tool.setuptools]
    include-package-data = true

[tool.ruff]
    line-length = 88

[tool.ruff.lint]
    ignore = [
        "D",       # docstrings
        "G004",    # logging-f-strings
        "PLR2004", # magic-value-comparison
        "S101",    # assert
        "S311",    # suspicious-non-cryptographic-random-usage
        "SLF001",  # private-member-access
    ]

[tool.uv.sources]
    snapcogs = { workspace = true }

[dependency-groups]
    dev = [
        "rich>=13.9.4",
    ]
